fix:积分累加使用int导致小数点后数据缺失

+++ specificApis/function.py
@@ -85,7 +85,7 @@ def getStudentData(studentId):
 
     durations = 0  # 总时长s
     for data in stuDatas:
-        points += int(data['points'])
+        points += float(data['points'])
         durations += int(data['duration'])
     student['number'] = len(stuDatas)  # 总次数
     student['durations'] = durations
